eepy

+++ view/js/comun.js
@@ -1,24 +1,31 @@
 // Recupera el nombre del usuario almacenado en localStorage
 var username = localStorage.getItem('usuario');
 var isAdmin = localStorage.getItem('is_admin');
+var userIp = localStorage.getItem('ipUser');
+
+console.log('Ip:', userIp);
 
 if (username) {
     fetch('../../controller/cCookieController.php')
         .then(response => response.text())
         .then(data => {
-            data = JSON.parse(data);
-            if (data.success) {
-                // Mostrar toast de éxito al cerrar sesión
-                toastr.warning(data.message);
-                localStorage.clear(); // Eliminar todos los datos del localStorage
-            } else {
-                toastr.error(data.message);
-            }
+            if (data.trim() !== '') {
+                data = JSON.parse(data);
+                if (data.success) {
+                    // Mostrar toast de éxito al cerrar sesión
+                    toastr.warning(data.message);
+                    localStorage.clear(); // Eliminar todos los datos del localStorage
+                } else {
+                    toastr.error(data.message);
+                }
 
-            // Opcional: Redireccionar o realizar otras acciones después de cerrar sesión
-            setTimeout(function () {
-                window.location.href = "../html/index.html";
-            }, 1000);
+                // Opcional: Redireccionar o realizar otras acciones después de cerrar sesión
+                setTimeout(function () {
+                    window.location.href = "../html/index.html";
+                }, 1000);
+            } else {
+                console.log('Hay cookies almacenadas');
+            }
         })
     $('#logReg').hide();
     if (isAdmin === "1") {
